Refactored script into jupyter notebook
Refactored the python script written by David Strauss into a Jupyter
notebook. Also made an effort to change the script to be easier to
maintain and to read:
- Reafactored the calculations for all of the variables into their own
  functions.
- Refactored the iterative loop into it's own function.
- Streamlined the creation of the DataFrame product.

It's not perfect, and I probably made a few mistakes along the way. My
main goal was to put a framework around the model that would make it
easier to maintain and fix in the future by isolating logic wherever possible.

The products of this model will need to be vigorously vetted to ensure
that they are correct. It is highly likely that during the translation
of the original model to python by David S, and then to Jupyter by me,
a mistake was made and the model broken.