Refactor after conversation with Max
Had a video call with Max to verify the fidelity of the py model to the
original excel model. Found a few inconsistencies.

Changed r0 to r. r0 is the initial value of r.

Added a couple of static variables, and changed
case_fatality_rate_hospitals_overwhelmed and
rolling_intervals_for_current_infected to be calculated from those
vairables at the start of the model execution. Max felt it was a more
faithful representation of reality.

Isolated the way we determine if a hospital is overloaded and the way we
calculate 'effective' r. Doesn't make a huge difference now, but will
make it easier to maintain and update in the future.