Refactored model to accept all model assumptions at runtime
Model assumptions and parameters were set as global variables. Now they
are passed into the model as a dict which is referenced as necessary by
the various functions.

This has the advantage of allowing us to define the model assumptions
and execute the model in the same cell, which is more conducive to an
interactive dashboard