Refactored model ingestion of timeseries
Changed the model so that it now  receives the timeseries data from the
model_parameters dict rather than retrieving it on it's own. This gives
the user greater control over the data being fed into the model

Updated the Datasets class to parse the dates into datetimes. Makes for
easier manipulation down the road.

Changed the way that the snapshots are created.