Refactored with Max and David's additions to the model
Added ability to artificially set the r value at specific dates to
simulate interventions.

Model now backfills data to Mar3 if it the records don't reach back that
far. The number of cases is set as 1/2 of the following day. As I write
this I think that needs to change. For another commit.

Model also cuts off the data on Mar3. If any data exists that precedes
that date, it filters it out.