Added model output snapsnots
The model in it's current state has been vetted for accuracy. To fully
leverage this moment, I have created a directory of snapshots which
reflect the output of the model.

A test class has been created to compare the output of the model in it's
current configuration to these outputs, so that if the output begins to
change we know about it. Should give us more confidence about the
model's accuracy in the future.

As I write this I realize that in order for this to work we need to
stabilize the inputs as well. That's next