Dataset class change
After reviewing the JHU data, I realized that we needed to adjust the
way the Datasets retrieve the data to allow for data of an unknown
structure.

The get_'____'_data functions are now expected to be overridden by the
child classes and will perform not only the retrieval but also the
formatting of the incoming data into the expected structure.