Restructuring dataset and model to setup parameters/data more declaritively.
The attempt here is to pull as much of the data setup into initialization steps.
This will make testing the classes much easier because it becomes much easier
to pass in scenarios that you want to test and no mocking is needed to wrangle
how that data gets loaded.

There is still some more separation that could happen.  The CovidDatasets class
currently loads the pandas data in the intialization function. This probably would
be better served to run intialization outside of the init function.