Fixed the missing backfill dates problem
Swapped the order of the cutoff and the backfill functions in the data
prep step.

The problem was being caused by holes in the data. If there is data on
feb 1, but nothing else until mar 10, then the backfill function will do
nothing because there's data prior to the cutoff date, then the cutoff
function will remove the feb data and the first date left will be the
10th. Now we cutoff first and backfill second.