Unify env vars in deploy workflow.
I mostly just wanted a single place where I can update the covid-data-model
branch that's being run against (I almost messed this up when trying to do a
test yesterday). But I took the opportunity to unify all the env vars that are
used repeatedly and/or are important enough to be top-level constants.