Update covid_tracking_source to check and scrub non-sense ICU values
Removes the 'current icu' data in the covid tracking dataset anytime it is greater than the 'current hospitalized' amount. Motivated by covid-tracking providing a new value for Delaware current ICU that is not consistent with the capacity and hospitalized data.

Also added an assert statement to check for this condition going forward once the Delaware problem is fixed upstream.