ICU Estimation from Cumulative ICU Data
Improve the estimation of current icu covid patients in the case where we are only provided with cumulative icu patients.

Sometimes covidtracking provides cumulative, but not current, data for hospitalization and icu. We then estimate the number of patients currently in the ICU. There are multiple things with this implementation that need to be improved, however they can be done in stages. The first stage is to standardize the calculation of ICU. Right now the estimate uses a variable drawn from a wide distribution (roughly 8.6 +/- 2.6 days). This introduces a large source of noise in the snapshot to snapshot calculations. In other places this code is usually called multiple times and then averaged, but for here it is just called once. Instead of bumping up the call/average, I just calculated the value from the means. The consistancy day-to-day is key, especially since this parameter isn't used in any monte-carlo modeling.

I left the hospitilization branch still drawing from the distribution, but averaging over 250 samples. Timeit shows an increase from 8.4ms to 27.2ms. If this becomes significant, we can refactor the params code and surface this constant.

Finally, our new linter just came online, so there will be misc line cleanups for a bit.