Refactored Current to Cumulative Outside Load Data
Extracted the logic to convert cumulative totals to an estimate of current patient population. Before this, it was added into a data loading function as an optional flag. That structure was too constraining to include required features of the model (i.e. requiring more than one value provided to extrapolate from).