Fix ICU Utilization Metric When Incorporating New 'Cumulative ICU' Datastream (#427)
Updated the 'estimate current number of hospital/icu beds occupied' by improving the model initialization and removing some unplanned sources of variability.

* ICU Estimation from Cumulative ICU Data

Improve the estimation of current icu covid patients in the case where we are only provided with cumulative icu patients.

Sometimes covidtracking provides cumulative, but not current, data for hospitalization and icu. We then estimate the number of patients currently in the ICU. There are multiple things with this implementation that need to be improved, however they can be done in stages. The first stage is to standardize the calculation of ICU. Right now the estimate uses a variable drawn from a wide distribution (roughly 8.6 +/- 2.6 days). This introduces a large source of noise in the snapshot to snapshot calculations. In other places this code is usually called multiple times and then averaged, but for here it is just called once. Instead of bumping up the call/average, I just calculated the value from the means. The consistancy day-to-day is key, especially since this parameter isn't used in any monte-carlo modeling.

I left the hospitilization branch still drawing from the distribution, but averaging over 250 samples. Timeit shows an increase from 8.4ms to 27.2ms. If this becomes significant, we can refactor the params code and surface this constant.

Finally, our new linter just came online, so there will be misc line cleanups for a bit.

* Parameter Scope Reduction for Temporary Test

The added parameter tripped up code elsewhere that has a script list of allowed keys. This constant is now local, but will be incorporated into the params later.

* Refactoring Current_To_Cumulative Code

Trying to isolate the code that converts current to cumulative from the code that just loads any data.

* Refactored Current to Cumulative Outside Load Data

Extracted the logic to convert cumulative totals to an estimate of current patient population. Before this, it was added into a data loading function as an optional flag. That structure was too constraining to include required features of the model (i.e. requiring more than one value provided to extrapolate from).

* Documentation and Reformatting