Bugfix: Build-all Passes Whitelist that is Ignored by WebUIDataAdaptorV1
Initial motivation was to abstract the hospitalization and icu conversion ratios so that they could be unit tested. In getting the __main__ loop running, there appears to be a mismatch between the whitelist passed by build-all, and the one used in the WebUIDataAdaptorV1. I have standardized this so that it uses the build-all whitelist.