Bugfix: Build-all Passes Whitelist that is Ignored by WebUIDataAdaptorV1 (#469)
* Bugfix: Build-all Passes Whitelist that is Ignored by WebUIDataAdaptorV1

Initial motivation was to abstract the hospitalization and icu conversion ratios so that they could be unit tested. In getting the __main__ loop running, there appears to be a mismatch between the whitelist passed by build-all, and the one used in the WebUIDataAdaptorV1. I have standardized this so that it uses the build-all whitelist.

* Keep County Whitelist Generation in Build-All

Build-all in the cli.py will do all filtering of possible counties to output and webui_data_adaptor_v1.py will run on all counties passed to it through generate_state.

* Building Snapshot

* Remove Deploy API Hook