Update Posteriors if Denominator is Zero (#486)
We now have an initial prior and a reinitializing prior for the rt code. They were originally the same (mode at 1.5, mean at 2.5). They are now separate, and we reinitialized at a lower value (mode=1, mean=2). This falls in the class of "a string of zero cases goes to r->1" and will be re-visited once Alex's changes land.

* Update Posteriors if Denominator is Zero

https://trello.com/c/KXNyFaYc/219-low-n-rt-posterior-reset-causes-red

* Trigger Snapshot with Previous Day Posterior

* Snapshot with More Centered Prior