Apply Shim to Cumulative Death s.t. Latest Observed Equals Latest Model (#502)
Our model tracks incident deaths. We present a cumulative deaths value in the FE in the table for different forecasted observations.

The model can accumulate tracking error for the cumulative deaths. In some cases this is significant enough to raise questions in the "estimated cumulative deaths". We are actively re-evaluating the forecasts and table, but for now, we apply a shim to correct for this tracking error. This will take the cumulative deaths timeseries and apply a constant value to shift the graph s.t. today's model value matches with today's observed.

https://trello.com/c/ZDNt3or6/247-shim-model-cumulative-deaths-to-observed

* Apply Shim to Cumulative Death s.t. Latest Observed Equals Latest Model

This code shims the cumulative deaths timeseries in the API so that the cumulative deaths from the model agrees with the cumulative deaths from outside observed sources.

* Updated as per PR Feedback

* Remove Trigger