GitHub Self-Hosted Runner Queue (#514)
Ensure a Snapshot is Sent to the Github Action Runner as a Single Unit of Work

Enforce a queue to the GitHub self-hosted runner that treats the creation of the entire snapshot as the job.

The previous version separated the workflow into multiple "jobs" that were linearly dependent on each other. That "jobs" abstraction is built to allow parallel processes, and if we used the GitHub hosted version it would run up to 20 jobs in parallel. However, the runner built for self-hosting prioritizes the jobs across snapshots in a way that delays the time to first snapshot completion (without changing the time to last snapshot completion). Instead of completing all the jobs of the first snapshot, it will cycle through each outstanding snapshot request calling the next available job for that snapshot.

It is better to have the early snapshot complete and unblock the person waiting for it.

This is solved by switching a given snapshot to a single job with multiple steps. The "steps" abstraction is the correct mental model that enforces sequential ordering (which we want anyways in our current implementations). So this will send all the commands to the runner as one packet.