Infer RT Improvements - Auto Sigma, Avoid Rounding (#505)
Changes to Rteff inference process that improve accuracy for low counts (changes to the gaussian sigma process, avoiding rounding of smoothed timeseries) and general interpretability of the results (removing deaths and hospitalizations from the composite Rt, increasing the smoothing window width for cases and introducing smoothing to the resulting Rt). The resulting Rt timeseries are much more accurate for low case counts and agree much more closely with leading sites such as rt.live and epiforecasts.io though there are still some differences (see below).

The logic for generating the gaussian process_matrix (which supports Rt changing over time by smoothing the prior distribution each day) is generalized to auto adjust the amount of smoothing depending on the recent (last 10 days or so) scale of counts (width of smoothing window scales with 1/sqrt(counts) to a maximum). This addresses a bias where new likelihood signals from low case counts (which are weaker having a broader distribution) are not able to move a narrow prior distribution causing Rt to significant lag (up to >30 days in some cases) even in the presence of a consistent signal in likelihood. An issue with symmetry of the smoothing window, that was already present but less frequently impactful due to the low value of sigma) was also addressed removing a bias towards Rt values in the middle, Rt = (10 + 0)/2 = 5, of the distribution range. These changes much improve accuracy for lower case counts (<200 impacting about half of all states and most counties) but reduce the amount of “smoothing” in the inference process allowing more existing noise to propagate into the result.

Additionally the process for generating the likelihood signal, which previously needed an integral input, was generalized (using interpolation) to accept float inputs which dramatically improves accuracy of Rt and reduces (shot) noise for very low case counts (<10).

In general the accuracy and consistent interpretability of Rt inference is also greatly impacted by returning a composite result that is the weighted average of Rt determined from timeseries for cases (what is expected) and deaths (generally not done). While there were good motivations for this, side effects include poor interpretability (peaks in new cases no longer align with Rt=1, …), poor consistency (sometimes deaths are not available or have much lower stats). Additionally deaths lag cases by many days so this reduces the usefulness of Rt as a leading indicator. For these reasons the composite Rt was changed to be derived only from cases (though the other contributions are still calculated so they can be used for other purposes).

Finally the smoothing applied to the case timeseries was improved to ensure that the window_size (14) and kernel_width (5) are compatible. The existing code has too narrow a window_size which clipped the gaussian kernel introducing sudden discontinuities in the slope of the smoothed signal. This was greatly reduced by increasing the window_size to 19. Smoothing was also introduced to the final Rt signal which is able to reduce noise (and improve confidence intervals) without introducing further artifacts into the process of the Rt inference process. This smoothing conserves the integral of Rt over time and hence conserves the overall scale change of the underlying case signal.

The end result of all these changes is an Rt signal that is much more accurate for low counts, much more interpretable, and much more consistent with results from other leading sites publishing Rt values. There are still the following issues to address:
1. There are no unit tests that quantitatively compare inferred Rt to what is known to be the correct value (in test data)
2. The case smoothing process is not consistent in time. At the start and end of a timeseries it “slows down” which forces Rt 50% from the true value towards 1
3. The testing volume adjustments made in the code do not seem to be as effective as those made by others (e.g. rt.live)
4. Our logic, based on the Bettencourt approach, is consider “old” now and leading sites are using more general and robust MCMC algorithms. But these are 100s of times slower and will probably never be viable to run at the county level.
5. Rt results are in “reported” time while many other sites show “onset” time so our results lag those by about 7 days. Whether or not to change this is a product question.