Remove Import Rename in CLI by Renaming Fun
As per PR#530, we are simplifying the Bettencourt Rt inference work. Part of that work was making a separate submodule. The cli api was inconsistent about calling this method (all the other ones prepend run_). This change fixes that inconsistency. It will break old api calls that directly called infer_rt (instead of run_infer_rt). The cli is almost exclusively used internally, so I think this change is worth the possible confusion, especially because it will fail loudly and quickly.