Refactor Infer_Rt Step 2 (#530)
This PR is the first part of a refactor for our rt inference code.

The infer_rt code now has its own submodule, and has been refactored to accept data agnostic to the aggregation level. This submission does not change any of the major internal components. It just structures the infer_rt code in its own submodule and pulls apart the RtInferenceEngine class into more manageable chunks. Now that Alex's significant improvements have landed, it's time to pay down that technical debt while allowing us to investigate new aggregation levels.

This change is in support of refactoring the inputs into RtInferenceEngine to be agnostic to aggregation level, which will allow for more manageable testing, maintenance, extensibility, and readability. Also, the tests for infer_rt are now under the same file in the test dir.

Outstanding Changes Remaining:
- A change to make the test correction more responsive to data from the last week was removed and will be put as a separate PR.
- Chris made some good recommendations about how to change the internal structure of the code, which will be a separate PR.
- Cli revamps to call a just the infer rt code was separated as a separate PR.
- The whitelist for pyseir and for infer_rt should be completely separated.


* initial commit from stash

* tail suppression addressed, compared with rt.live and inference unit tests working

* generate snapshots for this branch

* fixed breaks caused by merge from master

* first pass at modifying test volume adjustment - not working well yet

* Move infer_rt to Separate Submodule

* ready for Brett made sure all tests working

* Remove Hospitalization Data References from Infer RT

* Remove Cross Dependency on Pyseir Parameter Generation

Note: The serial period is now fixed in InferRtConstants instead of estimated from pyseir. I think it's reasonable to maintain these separately.

* Enforce RtInferenceEngine Accesses InferRtConstants Only Through Init

First step to isolate cross dependencies on the infer_rt script. Most of these should be class variables instead of instance variables, but that's for another time. Right now the goal is to make it clear that everything in the RtInferenceEngine is self contained.

* Extract Plotting Boilerplate to Separate File

* Cleanup for PR Submission

* Update deploy_api.yml

* Fix requirements vs requirements_test error

* Pull Data Loader out of RtInferenceEngine

Make the core Infer Rt Code Agnostic to the Aggregation Level.

* Remove Import Rename in CLI by Renaming Fun

As per PR#530, we are simplifying the Bettencourt Rt inference work. Part of that work was making a separate submodule. The cli api was inconsistent about calling this method (all the other ones prepend run_). This change fixes that inconsistency. It will break old api calls that directly called infer_rt (instead of run_infer_rt). The cli is almost exclusively used internally, so I think this change is worth the possible confusion, especially because it will fail loudly and quickly.

* Apply Filtering at the Data Loading Step

* Refactor Tests to Take Advantage of New Infer Rt Signature

The infer rt tests now use the new method of passing in a dataframe. All tests pass, but I think that's more a function of bad test coverage. Especially around initialization. A good opportunity to create more tests!

* Infer Rt Tests

* Creating Merged DataFrames

* initial work to get tests running correctly after the refactor and to get some of the charting back. More cleanup still needed

* rt inference tests running, figure_collector implemented and used by tests, charting of smoothed cases included again

* Adding test_results directory + a bit of test cleanup

* Removing Merge df Code to be Added as a Separate Branch

* disable data qa

* Tweaking deploy

* Remove space

* trying to manually pull lfs

* Revert Changes to Test Positivity Correction

Alex's Improvements to Test Positivity Correction Will be on a Separate PR

Co-authored-by: Alex MacAulay <macaulay.sandy@gmail.com>
Co-authored-by: Chris Kelly <ghop02@gmail.com>