merge latest and timeseries definitions (#585)
This PR is not expected to preserve all the existing behavior while making it easier to maintain `ALL_FIELDS_FEATURE_DEFINITION` and `ALL_TIMESERIES_FEATURE_DEFINITION`. They are almost identical and now only the differences between them are in the literal value for `ALL_FIELDS_FEATURE_DEFINITION`.

In a followup PR we may want to make the definitions closer to each other by removing elements from the second literal.