change indexing of summarize_timeseries_fields and add test (#606)
`dataset_summary_gen.py` depended on the order of the indexes set in `summarize_timeseries_fields` but it wasn't tested. This PR adds a test and changes the indexing to add only dates. My next PR adds another caller to dataset_summary_gen which can't as easily set the same index.