Separate Generation of Infection Rate Metric from PySEIR (#638)
Modify the infection rate calculations to:
- Use a region whitelist separate from the SEIR code
- Return the output at the end of the call in addition to internally saving an intermediate artifact to disk
- Save a table for infection rate metric keyed on region_id and date and stored in the snapshot artifact.

* Initial Commit to Separate Generation of Infection Rate Metric from PySEIR

* Use pipeline.Region Instance for Passing Fips Information

* Apply Recommended Changes from PR

* Make infer_rt.run_rt Always Return a DataFrame (Empty if Need Be)

* Update run.sh

* Added Generate Infection Rate Metric Tests

* run.sh Test

* Update run.sh

* Update Infer Rt to Catch Empty Fips Calculations

Fips of the form XX999 are placeholder fips that often don't have data. There are currently some assertions in the case loading code that require non empty response packets. This change catches those AssertionErrors and handles them by returning empty pd.DataFrames which are correctly handled downstream.

* Update infer_rt.py

* Updated Error and Test Messages

* Skip MA Counties but Add Test Coverage

This refactor will not fix the MA Counties (which are currently not provided due to NaN handling). There is a test to acknowledge this expected failure, and we will filter out upstream.

* Documentation Updates

* Update cli.py

* Update infer_rt_test.py