Aggregate to MSAs by default (#764)
## This PR

Changes combined datasets build to aggregate to MSAs by default. It increases the size of multiregion.csv from around 100MB to 120MB. The wide dates representation takes 10MB for the same data, but is not used yet.

## Tested

Ran a snapshot in this branch and compared it to one built with the same data in master. They look identical: https://covidactnow.org/internal/compare?left=1205&locations=2&metric=5&right=1207&sort=3