MultiRegionDataset combine for test positivity (#818)
This PR adds a new function for combining MultiRegionDataset and changes the existing test_positivity module to use it instead of its own method. It builds on https://github.com/covid-projections/covid-data-model/pull/817

This change makes it easier to create a set of methods, each producing timeseries with provenance information for subsets of the regions and then combine the results.

Tested by running `./run.py api generate-test-positivity --output-dir results/output-branch/` and similarly in master branch, then comparing the output files. The only difference is `variable` changed to `dataset`.