add _apply_scaling_factor and test (#813)
This PR builds on https://github.com/covid-projections/covid-data-model/pull/811 to add scaling of certain timeseries using certain static values.

Not hard-coding the fields may be making this code more flexible than we need, but does make writing small unit tests easier.

This development direction uses panda parallel operations pretty heavily to get the speed to aggregate all CBSA's in a few seconds on my laptop. This makes the code somewhat harder to read because each value and operation must handle multiple locations at the same time. For example, instead of multiplying by a scalar weight the code needs to multiple by a vector of weights indexed by location.

Alternative development directions that have code paths processing one location at a time are likely easier to read but harder to run. Panda's builtin aggregation functions are not flexible enough to do things like weighted averages. It'd be really nice to have a function that takes the unaggregated static + timeseries DataFrame for each aggregated region, and returns the static and timeseries of that one region. A simple test with `apply` suggests this will take ~30min to run with bare python. It is quite possible numba can bring that down to a few seconds, but I haven't tried. It'd be nice if we had an easy way to run this in the cloud (I'd like to try Dask/Prefect) but that is a much bigger task.