Move annotations into wide-dates file (#933)
This PR prepares for https://trello.com/c/lbiZpKwH/849-proposal-improve-the-provenance-link-so-that-we-can-show-a-source-urls-instead-of-a-data-class by moving all tags into the wide dates file. Before this change provenance was the only tag written to the wide-dates file. Reducing the number of files to represent a MultiRegionDataset will also simplify reading it directly from a git history.