read and transform parquet file during data update (#934)
This PR will help with https://trello.com/c/lbiZpKwH/849-proposal-improve-the-provenance-link-so-that-we-can-show-a-source-urls-instead-of-a-data-class because it lets us skip adding source_url in covid-data-public CSVs. It also simplifies the process of adding new fields in general to combined_datasets.

* Adds a new (and hacky) `CanScraperBase` that caches the CAN parquet file after it has been loaded
* Changes `CANScraperStateProviders`, `CDCTestingDataset` and `CDCVaccinesDataset` to depend on their scripts in the covid-data-public repo. After the scripts are turned off we can move the source to this repo
* Adds a pip dependency on python-snappy to fix  ..../fastparquet/compression.py line 144, in decompress_data RuntimeError: Decompression 'SNAPPY' not available. 

## Tested

In `main` ran `time ./run.py data update` and got
```
real	14m53.280s
user	10m39.442s
sys	3m46.207s
```

then checked out this branch, ran the same and got
```
real	9m43.823s
user	7m48.892s
sys	1m39.757s
```

with the same output in `data` except for the covid-data-model repo state.