Add source_url tag (#935)
This PR is a big part of https://trello.com/c/lbiZpKwH/849-proposal-improve-the-provenance-link-so-that-we-can-show-a-source-urls-instead-of-a-data-class

* Adds `class SourceUrl(TagInTimeseries)` and an property to get them in `OneRegionTimeseriesDataset`
* Changes `test_helper.TimeseriesLiteral` to a frozen dataclass with some type conversion in `__init__`
* Adds `dataclass_with_default_init`, backport of code in Python 3.9

Instead of doing type conversion in `test_helper.TimeseriesLiteral` i could leave the attributes as one of 3 types and added properties provenance_as_list and source_url_as_list. Either is fine but I had to pick one.