Move variables to a CanScraperBase attribute (#939)
This PR is part of https://trello.com/c/lbiZpKwH/849-proposal-improve-the-provenance-link-so-that-we-can-show-a-source-urls-instead-of-a-data-class

replace CanScraperBase.TRANSFORM_METHOD with VARIABLES and a method that can be overridden. This will make it easier to have query_multiple_variables return source_url values that can be added to the MultiRegionDataset.

## Tested

Ran `data update` and checked that identical output was produced.