source_url in API (#942)
This PR is part of https://trello.com/c/lbiZpKwH/849-proposal-improve-the-provenance-link-so-that-we-can-show-a-source-urls-instead-of-a-data-class

* Adds ￼source_url to API FieldAnnotations
* Adds annotations for vaccines which were in the API but never populated.
* Adds a tests to make sure all FieldAnnotations are populated

## Tested

Changed to branch `git checkout tgb-849-add-source-url`, ran `./run.py data update` then switched back to this branch with updated data. Then ran `python pyseir/cli.py build-all --generate-api-v2 -s FL --fips 12011` and checked with `jq . output/county/12011.timeseries.json` that some URLs were added.