Buckets from CanScraperLoader (#987)
This PR is part of https://trello.com/c/lvRFxC5y/932-get-dei-vaccination-data-into-api

* Changes CanScraperLoader to return rows of observations identified by location_id. It returns all buckets but getting buckets that are not "all" to work further downstream is a WIP
* Adds a check that `location_id` in parquet file matches what is computed from FIPS. This fails right now so added work around.
* Updates `test_helpers.assert_dataset_like` to compare `timeseries_bucketed` instead of `timeseries`

## Tested

Ran `data update` many many times, fixing various problems until it was identical. This is the expected state because stages of the pipeline after loading the data copy `timeseries` instead of `timeseries_bucketed`, dropping all timeseries that are not in bucket "all".