Keep least stale vaccine percentage (#1090)
This PR is part of https://trello.com/c/tNlZtGp6/1279-make-our-vaccine-backfill-code-smarter

* Changes `vaccine_backfills.derive_vaccine_pct` to replace existing percentage metrics when the percentage derived from the people count has a real value on a more recent date

## Tested

I hacked in a change to write a wide-dates csv of the dataset before and after this change.
For counties in MD this adds a few days of data to `vaccinations_completed_pct` , for counties in ME this adds about 20 days of data to `vaccinations_initiated_pct`. It looks like the corresponding CBSAs had similar changes.