Estimate 1st dose data if it's been missing recently. (#1155)
Previously we wouldn't estimate 1st dose data from completed if there was any
1st dose data at all.

With this change, we look back 16 days and if there's no 1st dose data in that
span, then we estimate 1st dose from completed. This is necessary since some
of the new CDC timeseries data has stale 1st dose data for some reason and we
want to go ahead and use estimated data up to present so we can show it
on the website.