Update README docs and fix states/fips options to data update --no-refresh-datasets. (#1271)
Updates README.md and SETUP.md to remove outdated information and include better instructions for running (parts of) the pipeline locally. It's not perfect, but hopefully it's a step in the right direction.

Also made a drive-by improvement to make ./run.py data update respect the --state and --fips arguments when you pas --no-refresh-datasets.