Allow Unknown DNC in Community Level (#1367)
* Allow Unknown DNC in Community Level

* Updating saved datasets at Fri Apr 14 21:32:45 UTC 2023

* Pushing Broken Test

This test was passing incorrectly because the assert was wrong. Now it should be failing, which represents the broken state of this PR.

* add null checking for metrics in community level, remove forward filling

* Updated Tests to Fail in More Interesting Ways

Adds test for default to 0 for stale DNC after 14 days.

We've had some off-by-1 uncertainty in what "14 days lookback before blocking looks" really means. Here's an updated test, that currently fails, that captures the expected behavior.

I propose that we either change the code to handle this correctly, or update our docs to say 15 whenever they used to say 14.

* fix off by one error

---------

Co-authored-by: GitHub Action <action@github.com>
Co-authored-by: Sean McClure <sean.mcclure17@gmail.com>